Fixed prerendering error with suspense boundaries
